Added support for non-windows linebreaks
Most of the time you just get \n, even if you submit text from a Windows PC it seems to end up as \n instead of \r\n for me.

But this change will permit both. I double-checked that it wouldn't replace the \n on its own first, and it doesn't. It checks \r\n then \n